renamed project-list to item-list
